move swap tests into a separate test file, add missing data handle
